image_generation_service/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-aiplatform
vertexai
gunicorn